refactor(cli): focus terminal title lifecycle
